+++ darto_validator/lib/darto_validator.dart
@@ -1,5 +1,6 @@
 library darto_validator;
 
 export 'package:zard/zard.dart' show Schema, ZardResult, ZardError, z;
+export 'src/error_bridge.dart';
 export 'src/openapi_schema.dart';
 export 'src/z_validator.dart';